requirements.txt
# AIOS CI/validation Python dependencies
# Used by smoke tests, schema validation scripts, and build tooling

PyYAML
jsonschema
